farm_packt/beanie_backend/requirements.txt
fastapi
fastapi_cors
beanie
bcrypt
cloudinary
uvicorn
pydantic-settings 
PyJWT
python-multipart
openai
resend
